Refactored namespace of resource classes
