+++ Pirat.DatabaseTests/DemandServiceTest.cs
@@ -10,6 +10,7 @@ using Pirat.Examples.TestExamples;
 using Pirat.Exceptions;
 using Pirat.Model;
 using Pirat.Model.Entity;
+using Pirat.Model.Entity.Resource.Common;
 using Pirat.Services;
 using Pirat.Services.Resource;
 using Xunit;
