+++ Pirat/Model/Entity/Resource/Common/Findable.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Pirat.DatabaseContext;
 
-namespace Pirat.Model.Entity
+namespace Pirat.Model.Entity.Resource.Common
 {
     /**
      * For an entity implementing Findable queries by unique id can be made to find the entry in the database
