+++ Pirat/Model/Entity/Resource/Demand/DemandEntity.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Pirat.DatabaseContext;
 using Pirat.Model.Api.Resource;
+using Pirat.Model.Entity.Resource.Common;
 
 namespace Pirat.Model.Entity.Resource.Demand
 {
