+++ Pirat/Model/Entity/Resource/Demand/ItemDemandEntity.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Pirat.Model.Entity.Resource.Demand
 {
