+++ Pirat/Services/Helper/AddressMaker/AddressMaker.cs
@@ -6,6 +6,7 @@ using Newtonsoft.Json.Linq;
 using Pirat.Codes;
 using Pirat.Exceptions;
 using Pirat.Model;
+using Pirat.Model.Entity.Resource.Common;
 
 namespace Pirat.Services.Helper.AddressMaker
 {
