+++ Pirat/Services/Mail/SubscriptionService.cs
@@ -7,6 +7,7 @@ using Pirat.DatabaseContext;
 using Pirat.Model;
 using Pirat.Model.Api.Resource;
 using Pirat.Model.Entity;
+using Pirat.Model.Entity.Resource.Common;
 using Pirat.Model.Entity.Resource.Stock;
 
 namespace Pirat.Services.Mail
