+++ Pirat/Services/Resource/ResourceStockQueryService.cs
@@ -11,6 +11,7 @@ using Pirat.Exceptions;
 using Pirat.Model;
 using Pirat.Model.Api.Resource;
 using Pirat.Model.Entity;
+using Pirat.Model.Entity.Resource.Common;
 using Pirat.Model.Entity.Resource.Stock;
 using Pirat.Services.Helper;
